Removed `DeveloperDeviceDetailsSettingsSelectionViewModel`
